1.8.3

+++ polyvLiveStreamerScene/src/main/java/com/easefun/polyv/livestreamer/modules/document/popuplist/PLVLSPptListLayout.java
@@ -47,6 +47,8 @@ import com.easefun.polyv.livestreamer.modules.document.popuplist.enums.PLVLSPptV
 import com.easefun.polyv.livestreamer.modules.document.popuplist.holder.PLVLSPptListViewHolder;
 import com.easefun.polyv.livestreamer.modules.document.popuplist.vo.PLVLSPptVO;
 import com.easefun.polyv.livestreamer.modules.document.popuplist.widget.PLVLSDocumentDeleteArrow;
+import com.plv.livescenes.access.PLVUserAbility;
+import com.plv.livescenes.access.PLVUserAbilityManager;
 import com.plv.thirdpart.blankj.utilcode.util.ConvertUtils;
 import com.plv.thirdpart.blankj.utilcode.util.SPUtils;
 import com.plv.thirdpart.blankj.utilcode.util.ScreenUtils;
@@ -154,6 +156,9 @@ public class PLVLSPptListLayout extends FrameLayout {
     // PPT文档上传监听接口实现
     private OnPLVSDocumentUploadListener documentUploadListener;
 
+    @Nullable
+    private PLVUserAbilityManager.OnUserAbilityChangedListener onUserAbilityChangeCallback;
+
     // </editor-fold>
 
     // <editor-fold defaultstate="collapsed" desc="构造方法">
@@ -188,6 +193,7 @@ public class PLVLSPptListLayout extends FrameLayout {
         PLVBlurUtils.initBlurView(plvlsBlurView);
 
         initMvpView();
+        initOnUserAbilityChangeListener();
     }
 
     private void findView() {
@@ -225,6 +231,9 @@ public class PLVLSPptListLayout extends FrameLayout {
         pptListAdapter.setOnPptItemClickListener(new PLVLSPptListViewHolder.OnPptItemClickListener() {
             @Override
             public void onClick(int id) {
+                if (!checkHasDocumentPermissionOrToast(PLVUserAbility.STREAMER_DOCUMENT_ALLOW_OPEN_PPT)) {
+                    return;
+                }
                 if (pptListAdapter.getRealViewType() == PLVLSPptViewType.COVER) {
                     // 显示为PPT文档列表
                     if (currentAutoId == id) {
@@ -261,6 +270,9 @@ public class PLVLSPptListLayout extends FrameLayout {
                                     .setRightBtnListener(new OnClickListener() {
                                         @Override
                                         public void onClick(View v) {
+                                            if (!checkHasDocumentPermissionOrToast(PLVUserAbility.STREAMER_DOCUMENT_ALLOW_DELETE_PPT)) {
+                                                return;
+                                            }
                                             PLVDocumentPresenter.getInstance().deleteDocument(fileId);
                                             documentDeleteConfirmDialog.hide();
                                         }
@@ -332,6 +344,9 @@ public class PLVLSPptListLayout extends FrameLayout {
         plvlsDocumentRefreshTv.setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View v) {
+                if (!checkHasDocumentPermissionOrToast(PLVUserAbility.STREAMER_DOCUMENT_ALLOW_PULL_PPT)) {
+                    return;
+                }
                 PLVDocumentPresenter.getInstance().requestGetPptCoverList(true);
             }
         });
@@ -385,6 +400,9 @@ public class PLVLSPptListLayout extends FrameLayout {
 
             @Override
             public boolean requestSelectUploadFileConvertType(final Uri fileUri) {
+                if (!checkHasDocumentPermissionOrToast(PLVUserAbility.STREAMER_DOCUMENT_ALLOW_UPLOAD_PPT)) {
+                    return true;
+                }
                 if (fileUri == null) {
                     Log.w(TAG, "file uri is null.");
                     PLVToast.Builder.context(getContext())
@@ -454,6 +472,13 @@ public class PLVLSPptListLayout extends FrameLayout {
                         .setRightBtnListener(new OnClickListener() {
                             @Override
                             public void onClick(View v) {
+                                if (!checkHasDocumentPermissionOrToast(PLVUserAbility.STREAMER_DOCUMENT_ALLOW_UPLOAD_PPT)) {
+                                    // 没有对应权限，取消上传
+                                    PLVDocumentPresenter.getInstance().removeUploadCache(cacheVOS);
+                                    documentUploadAgainConfirmDialog.hide();
+                                    return;
+                                }
+
                                 for (PLVPptUploadLocalCacheVO localCacheVO : cacheVOS) {
                                     File file = new File(localCacheVO.getFilePath());
                                     if (!file.exists()) {
@@ -503,6 +528,22 @@ public class PLVLSPptListLayout extends FrameLayout {
         };
 
         PLVDocumentPresenter.getInstance().registerView(mvpView);
+    }
+
+    /**
+     * 初始化用户角色能力变化监听
+     */
+    private void initOnUserAbilityChangeListener() {
+        this.onUserAbilityChangeCallback = new PLVUserAbilityManager.OnUserAbilityChangedListener() {
+            @Override
+            public void onUserAbilitiesChanged(@NonNull List<PLVUserAbility> addedAbilities, @NonNull List<PLVUserAbility> removedAbilities) {
+                if (PLVUserAbilityManager.myAbility().notHasAbility(PLVUserAbility.STREAMER_DOCUMENT_ALLOW_OPEN_PPT)) {
+                    close();
+                }
+            }
+        };
+
+        PLVUserAbilityManager.myAbility().addUserAbilityChangeListener(onUserAbilityChangeCallback);
     }
 
     // </editor-fold>
@@ -712,7 +753,10 @@ public class PLVLSPptListLayout extends FrameLayout {
      * 打开弹层
      */
     public void open(boolean refresh) {
-        if(refresh){
+        if (!checkHasDocumentPermissionOrToast(PLVUserAbility.STREAMER_DOCUMENT_ALLOW_OPEN_PPT)) {
+            return;
+        }
+        if (refresh) {
             PLVDocumentPresenter.getInstance().requestGetPptCoverList(true);
         }
         final int landscapeHeight = Math.min(ScreenUtils.getScreenWidth(), ScreenUtils.getScreenHeight());
@@ -787,6 +831,7 @@ public class PLVLSPptListLayout extends FrameLayout {
      * 销毁方法
      */
     public void destroy() {
+        onUserAbilityChangeCallback = null;
         stopUpdateBlurViewTimer();
     }
 
@@ -880,6 +925,16 @@ public class PLVLSPptListLayout extends FrameLayout {
                 }
             }, 3000);
         }
+    }
+
+    private boolean checkHasDocumentPermissionOrToast(PLVUserAbility documentAbility) {
+        if (PLVUserAbilityManager.myAbility().notHasAbility(documentAbility)) {
+            PLVToast.Builder.context(getContext())
+                    .setText(R.string.plvls_document_usage_not_permeitted)
+                    .show();
+            return false;
+        }
+        return true;
     }
 
     // </editor-fold>
